🚚 Move get/create/delete token queries/mutations/fragments into admin
