+++ src/admin/fragments.js
@@ -1,0 +1,14 @@
+import gql from 'graphql-tag';
+
+export const TOKEN_FIELDS = gql`
+  fragment TokenFields on DevDownloadTokenNode {
+    id
+    name
+    token
+    createdAt
+    creator {
+      username
+      picture
+    }
+  }
+`;
