+++ src/admin/queries.js
@@ -1,0 +1,16 @@
+import gql from 'graphql-tag';
+import {TOKEN_FIELDS} from './fragments';
+
+// Query to get developer tokens
+export const GET_DEV_TOKENS = gql`
+  query DevTokens {
+    allDevTokens {
+      edges {
+        node {
+          ...TokenFields
+        }
+      }
+    }
+  }
+  ${TOKEN_FIELDS}
+`;
