+++ src/admin/views/TokensListView.js
@@ -11,8 +11,8 @@ import {
   Message,
   Segment,
 } from 'semantic-ui-react';
-import {GET_DEV_TOKENS} from '../../state/queries';
-import {CREATE_DEV_TOKEN, DELETE_DEV_TOKEN} from '../../state/mutations';
+import {GET_DEV_TOKENS} from '../queries';
+import {CREATE_DEV_TOKEN, DELETE_DEV_TOKEN} from '../mutations';
 import {TokenList} from '../components/TokenList';
 import NewTokenForm from '../forms/NewTokenForm';
